# openapi.yaml
swagger: "2.0"
info:
  description: "API reference for the Freight Trust dApp which allows recording, transfering and signing of EDI-based documents such as bills of lading or warehouse receipts."
  version: "1.0.0"
  title: "Freight Trust - TrueNFT dApp API"
  termsOfService: "https://github.com/freight-trust/legal/blob/master/src/terms-of-service.md"
  contact:
    email: "corporate@freighttrust.com"
  license:
    name: "Mozzila Public License 2.0"
    url: "https://mozilla.org/MPL/2.0/"
host: "api.truenft.com"
basePath: "/"
...
schemes:
  - "https"
paths:
...
  /proxy/getTarget/:
    get:
      tags:
        - "Proxy and Upgrade"
      summary: "Gets address of Function"
      description: "Gets address that is delegate called to run a function's logic."
      operationId: "getTarget"
      parameters:
        - name: "id"
          in: "query"
          description: "The function selector the return the target of"
          required: true
          type: "string"
          maxLength: 10
          minLength: 10
      produces:
        - "application/xml"
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
...
  /proxy/setTarget/:
    post:
      tags:
        - "Proxy and Upgrade"
      summary: "Sets the address of Function"
      description: "Sets address that is delegate called to run a function's logic. This function must be called with an signed nonce from the Freight Trust account in the https body."
      operationId: "setTarget"
      parameters:
        - name: "id"
          in: "query"
          description: "Function selector whose target to change."
          required: true
          type: "string"
          minLength: 10
          maxLength: 10
        - name: "target"
          in: "query"
          description: "The address to set the function too"
          required: true
          type: "string"
          minLength: 42
          maxLength: 42
        - in: "body"
          name: "auth"
          description: "The Freight Trust signature of the nonce"
          required: true
          schema:
            $ref: "#/definitions/auth"
      produces:
        - "application/xml"
        - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/changeMaster/:
    post:
      tags:
        - "Proxy and Upgrade"
      summary: "Sets the address of Function"
      description: "Sets address that is delegate called to run a function's logic. This function must be called with an signed nonce from the Freight Trust account in the https body."
      operationId: "changeMaster"
      parameters:
        - name: "newMaster"
          in: "query"
          description: "The new master address"
          required: true
          type: "string"
          minLength: 42
          maxLength: 42
        - in: "body"
          name: "auth"
          description: "The Freight Trust signature of the nonce"
          required: true
          schema:
            $ref: "#/definitions/auth"
      produces:
        - "application/xml"
        - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/changeController/:
    post:
      tags:
        - "Proxy and Upgrade"
      summary: "Sets the the new controller address"
      description: "Sets the new controller address."
      operationId: "changeController"
      parameters:
        - name: "newController"
          in: "query"
          description: "The new controller address"
          required: true
          type: "string"
          minLength: 42
          maxLength: 42
        - in: "body"
          name: "auth"
          description: "The Freight Trust signature of the nonce"
          required: true
          schema:
            $ref: "#/definitions/auth"
      produces:
        - "application/xml"
        - "application/json"
      responses:
        200:
          description: "Successfuly Changed"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
  /proxy/pause:
    post:
      tags:
        - "Proxy and Upgrade"
      summary: "Pauses the execution"
      description: "Calls the pause method in the universal proxy, blocking any state changes while Freight Trust's Network Operations Group updates."
      operationId: "pause"
      parameters:
        - in: "body"
          name: "auth"
          description: "The Freight Trust signature of the nonce"
          required: true
          schema:
            $ref: "#/definitions/auth"
      produces:
        - "application/xml"
        - "application/json"
      responses:
        200:
          description: "Successfuly Paused"
          schema:
            $ref: "#/definitions/ApiResponse"
        425:
          description: "Invalid Input"
definitions:
  auth:
    type: "object"
    required:
      - "signature"
    properties:
      signature:
        type: "string"
        minLength: 132
        maxLength: 132
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...